feat: add alert dialogs for coupon elimination and event elimination

+++ app/src/main/java/com/example/unieventos/ui/screens/admin/AdminEventDetailScreen.kt
@@ -1,20 +1,17 @@
 package com.example.unieventos.ui.screens.admin
 
+import android.widget.Toast
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
-import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.filled.CalendarMonth
-import androidx.compose.material.icons.filled.Place
-import androidx.compose.material3.Icon
-import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.AlertDialog
 import androidx.compose.material3.Scaffold
 import androidx.compose.material3.Text
+import androidx.compose.material3.TextButton
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
@@ -22,15 +19,11 @@ import androidx.compose.runtime.saveable.rememberSaveable
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.layout.ContentScale
 import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.unit.dp
-import coil.compose.AsyncImage
-import coil.request.ImageRequest
 import com.example.unieventos.R
 import com.example.unieventos.enums.DateError
-import com.example.unieventos.ui.components.events.EventDetailItem
 import com.example.unieventos.ui.components.events.EventForm
 import com.example.unieventos.ui.components.utils.CustomTopAppBar
 import com.example.unieventos.ui.components.utils.PrimaryButton
@@ -38,7 +31,7 @@ import com.example.unieventos.ui.components.utils.SecondaryButton
 import com.example.unieventos.viewmodel.EventsViewModel
 
 @Composable
-fun EventDetailScreen(
+fun AdminEventDetailScreen(
     eventsViewModel: EventsViewModel,
     eventId: Int,
     onBack: () -> Unit
@@ -48,6 +41,8 @@ fun EventDetailScreen(
     if (event == null) {
         return
     }
+
+    val context = LocalContext.current
 
     var name by rememberSaveable { mutableStateOf(event.name) }
     var city by rememberSaveable { mutableStateOf(event.city) }
@@ -59,6 +54,8 @@ fun EventDetailScreen(
     var date by rememberSaveable { mutableStateOf(event.date) }
     var isDatePicked by rememberSaveable { mutableStateOf(false) }
     var dateError by rememberSaveable { mutableStateOf(DateError.NONE) }
+
+    var showDialog by rememberSaveable { mutableStateOf(false) }
 
     Scaffold(
         topBar = {
@@ -134,12 +131,50 @@ fun EventDetailScreen(
                 SecondaryButton(
                     text = stringResource(id = R.string.delete_event_btn),
                     onClick = {
-                        eventsViewModel.deleteEvent(event)
-                        onBack()
+                        showDialog = true
                     }
                 )
             }
         }
+    }
+
+    if (showDialog) {
+        AlertDialog(
+            onDismissRequest = {
+                showDialog = false
+            },
+            confirmButton = {
+                TextButton(
+                    onClick = {
+                        eventsViewModel.deleteEvent(event)
+                        showDialog = false
+                        Toast.makeText(
+                            context,
+                            R.string.event_deleted,
+                            Toast.LENGTH_SHORT
+                        ).show()
+                        onBack()
+                    }
+                ) {
+                    Text(stringResource(id = R.string.confirm_btn))
+                }
+            },
+            dismissButton = {
+                TextButton(
+                    onClick = {
+                        showDialog = false
+                    }
+                ) {
+                    Text(stringResource(id = R.string.cancel_btn))
+                }
+            },
+            title = {
+                Text(stringResource(id = R.string.delete_event_title))
+            },
+            text = {
+                Text(stringResource(id = R.string.delete_event_msg))
+            }
+        )
     }
 }
