+++ app/src/main/java/com/example/unieventos/ui/screens/admin/CouponDetailScreen.kt
@@ -1,5 +1,6 @@
 package com.example.unieventos.ui.screens.admin
 
+import android.widget.Toast
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
@@ -8,22 +9,20 @@ import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
+import androidx.compose.material3.AlertDialog
 import androidx.compose.material3.Scaffold
 import androidx.compose.material3.Text
+import androidx.compose.material3.TextButton
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableIntStateOf
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.saveable.rememberSaveable
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.layout.ContentScale
 import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.unit.dp
-import coil.compose.AsyncImage
-import coil.request.ImageRequest
 import com.example.unieventos.R
 import com.example.unieventos.enums.CouponCodeError
 import com.example.unieventos.enums.CouponNameError
@@ -46,6 +45,8 @@ fun CouponDetailScreen(
         return
     }
 
+    val context = LocalContext.current
+
     var name by rememberSaveable { mutableStateOf(coupon.name) }
     var nameError by rememberSaveable { mutableStateOf(CouponNameError.NONE) }
     var code by rememberSaveable { mutableStateOf(coupon.code) }
@@ -55,6 +56,8 @@ fun CouponDetailScreen(
     var expirationDate by rememberSaveable { mutableStateOf(coupon.expirationDate) }
     var isDatePicked by rememberSaveable { mutableStateOf(false) }
     var dateError by rememberSaveable { mutableStateOf(DateError.NONE) }
+
+    var showDialog by rememberSaveable { mutableStateOf(false) }
 
     Scaffold(
         topBar = {
@@ -136,14 +139,51 @@ fun CouponDetailScreen(
             ) {
                 SecondaryButton(
                     text = stringResource(id = R.string.delete_coupon_btn),
-                    enabled = true,
                     modifier = Modifier.fillMaxWidth(),
                     onClick = {
-                        couponsViewModel.deleteCoupon(coupon)
-                        onBack()
+                        showDialog = true
                     }
                 )
             }
         }
+    }
+
+    if (showDialog) {
+        AlertDialog(
+            onDismissRequest = {
+                showDialog = false
+            },
+            confirmButton = {
+                TextButton(
+                    onClick = {
+                        couponsViewModel.deleteCoupon(coupon)
+                        showDialog = false
+                        Toast.makeText(
+                            context,
+                            context.getString(R.string.coupon_deleted),
+                            Toast.LENGTH_SHORT
+                        ).show()
+                        onBack()
+                    }
+                ) {
+                    Text(stringResource(id = R.string.confirm_btn))
+                }
+            },
+            dismissButton = {
+                TextButton(
+                    onClick = {
+                        showDialog = false
+                    }
+                ) {
+                    Text(stringResource(id = R.string.dismiss_btn))
+                }
+            },
+            title = {
+                Text(stringResource(id = R.string.delete_coupon_title))
+            },
+            text = {
+                Text(stringResource(id = R.string.delete_coupon_msg))
+            }
+        )
     }
 }
